Echo test servlet will now run even if it can't find the spool directory (it logs a warning and spools to RAM instead).  compile.test now uses -source 1.5 -target 1.5.
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@16601 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ tests/com/l7tech/test/servlets/EchoServlet.java
@@ -6,33 +6,53 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.*;
+import java.util.logging.Logger;
+import java.util.logging.Level;
 
 /**
  * Disk-based echo servlet.  Spools entire post into a temp file; then reads it back and returns it.
+ * If it can't find a disk-based spoolDir, uses a memory-based spool instead.
  */
 public class EchoServlet extends HttpServlet {
+    protected static final Logger logger = Logger.getLogger(EchoServlet.class.getName());
+
     File spoolDir;
 
     public void init(ServletConfig config) throws ServletException {
         String spoolPath = config.getInitParameter("spoolPath");
         spoolDir = new File(spoolPath);
-        if (!(spoolPath != null && spoolPath.length() > 0 && spoolDir.exists() && spoolDir.isDirectory() && spoolDir.canWrite() && spoolDir.canRead()))
-            throw new ServletException("spoolPath is not a writable directory");
+        if (!(spoolPath != null && spoolPath.length() > 0 && spoolDir.exists() && spoolDir.isDirectory() && spoolDir.canWrite() && spoolDir.canRead())) {
+            logger.log(Level.WARNING, "spoolPath not found or not a writable directory - requests will be limited to what can fit into memory");
+            spoolDir = null;
+        }
+    }
+
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        resp.setContentType("text/plain");
+        resp.setStatus(200);
+        PrintStream out = new PrintStream(resp.getOutputStream());
+        out.println("Ready to echo POST requests to responses at this URL.");
+        if (spoolDir == null) {
+            out.println("spool directory missing, not configured, or not writable -- using memory spooling instead");
+        } else {
+            out.println("spooling files to " + spoolDir.getAbsolutePath());
+        }
     }
 
     protected void doPost(HttpServletRequest hreq, HttpServletResponse hresp) throws ServletException, IOException {
         String ctype = hreq.getContentType();
 
-        File file = File.createTempFile("echo", null, spoolDir);
-        FileOutputStream fos = null;
-        FileInputStream fis = null;
+        File file = spoolDir == null ? null : File.createTempFile("echo", null, spoolDir);
+        ByteArrayOutputStream baos = null;
+        OutputStream fos = null;
+        InputStream fis = null;
         try {
-            fos = new FileOutputStream(file);
+            fos = file == null ? (baos = new ByteArrayOutputStream()) : new FileOutputStream(file);
             copyStream(hreq.getInputStream(), fos);
             fos.close();
             fos = null;
 
-            fis = new FileInputStream(file);
+            fis = baos != null ? new ByteArrayInputStream(baos.toByteArray()) : new FileInputStream(file);
             hresp.setStatus(200);
             hresp.setContentType(ctype);
             copyStream(fis, hresp.getOutputStream());
@@ -41,7 +61,7 @@ public class EchoServlet extends HttpServlet {
         } finally {
             if (fos != null) try { fos.close(); } catch (IOException e) { /* ignore */ }
             if (fis != null) try { fis.close(); } catch (IOException e) { /* ignore */ }
-            file.delete();
+            if (file != null) file.delete();
         }
     }
